refactor: separated responsibility in application for add testings and best pratices
