+++ src/entities/Contact.ts
@@ -1,0 +1,8 @@
+export interface Contact {
+    id: string
+    name: string
+    email: string
+    subscribed: boolean
+    createdAt?: Date
+    updatedAt?: Date
+}
