+++ src/index.ts
@@ -1,7 +1,8 @@
 import multipart from '@fastify/multipart'
 import { fastify } from 'fastify'
-import { appRoutes } from 'routes'
-import { env } from './env'
+
+import { env } from '@/config/env'
+import { appRoutes } from '@/http/routes'
 
 const app = fastify()
 app.register(multipart)
