+++ src/lib/prisma-client.ts
@@ -1,0 +1,6 @@
+import { PrismaClient } from '@prisma/client'
+
+//adiciona as query no terminal
+export const prisma = new PrismaClient({
+    log: ['query'],
+})
